+++ ricaun.AppBundleTool/AppBundle/AppBundleInfoExtension.cs
@@ -33,9 +33,9 @@ namespace ricaun.AppBundleTool.AppBundle
             table.Columns.Add("Folder", typeof(string));
             table.Columns.Add("Bundle", typeof(string));
             table.Columns.Add("AppName", typeof(string));
+            table.Columns.Add("AppProduct", typeof(string));
             if (detail)
             {
-                table.Columns.Add("AppProduct", typeof(string));
                 table.Columns.Add("AppDescription", typeof(string));
             }
             table.Columns.Add("Access", typeof(string));
@@ -46,15 +46,78 @@ namespace ricaun.AppBundleTool.AppBundle
                 row["Folder"] = appBundleInfo.AppBundleFolder;
                 row["Bundle"] = appBundleInfo.Name;
                 row["AppName"] = appBundleInfo.ApplicationPackage?.Name ?? string.Empty;
+                row["AppProduct"] = appBundleInfo.ApplicationPackage?.AutodeskProduct ?? string.Empty;
                 if (detail)
                 {
-                    row["AppProduct"] = appBundleInfo.ApplicationPackage?.AutodeskProduct ?? string.Empty;
                     row["AppDescription"] = appBundleInfo.ApplicationPackage?.Description ?? string.Empty;
                 }
                 row["Access"] = appBundleInfo.AppBundleAccess;
+                //if (appBundleInfo.AppBundleAccess == AppBundleAccess.Allow)
+                //{
+                //    //row["Access"] = $"{GREEN}{appBundleInfo.AppBundleAccess}{NORMAL}";
+                //}
+                //else if (appBundleInfo.AppBundleAccess == AppBundleAccess.Admin)
+                //{
+                //    row["Access"] = $"{RED}{appBundleInfo.AppBundleAccess}{NORMAL}";
+                //}
                 table.Rows.Add(row);
             }
             return table;
+        }
+
+        public static DataTable ToDataTable(this AppBundleInfo appBundleInfo, bool detail = false)
+        {
+            if (appBundleInfo == null) return null;
+
+            var table = new DataTable("AppBundle");
+            table.CreateDataColumns();
+
+            table.DataRow("Bundle", appBundleInfo.Name);
+            table.DataRow("AppName", appBundleInfo.ApplicationPackage?.Name ?? string.Empty);
+            table.DataRow("AppVersion", appBundleInfo.ApplicationPackage?.AppVersion ?? string.Empty);
+            table.DataRow("AppProduct", appBundleInfo.ApplicationPackage?.AutodeskProduct ?? string.Empty);
+            if (detail)
+                table.DataRow("AppDescription", appBundleInfo.ApplicationPackage?.Description ?? string.Empty);
+            table.DataRow("AppProductType", appBundleInfo.ApplicationPackage?.ProductType ?? string.Empty);
+            table.DataRow("AppProductCode", appBundleInfo.ApplicationPackage?.ProductCode ?? string.Empty);
+            table.DataRow("AppCompanyName", appBundleInfo.ApplicationPackage?.CompanyDetails?.Name);
+            table.DataRow("PathBundle", appBundleInfo.PathBundle);
+            table.DataRow("Access", appBundleInfo.AppBundleAccess);
+
+            if (detail)
+            {
+                foreach (var component in appBundleInfo.ApplicationPackage?.Components)
+                {
+                    var componentText = component.RuntimeRequirements.Platform + " " +
+                                        component.RuntimeRequirements.OS + " " +
+                                        component.RuntimeRequirements.SeriesMin + " " +
+                                        component.RuntimeRequirements.SeriesMax;
+                    table.DataRow("Component", componentText);
+                    foreach (var componentEntry in component.ComponentEntry)
+                    {
+                        table.DataRow("ComponentEntry.AppName", componentEntry.AppName);
+                        table.DataRow("ComponentEntry.ModuleName", componentEntry.ModuleName);
+                    }
+                }
+            }
+
+            return table;
+        }
+
+        private static DataTable CreateDataColumns(this DataTable table)
+        {
+            table.Columns.Add("Name", typeof(string));
+            table.Columns.Add("Value", typeof(string));
+            return table;
+        }
+
+        private static DataRow DataRow(this DataTable table, string name, object value)
+        {
+            var row = table.NewRow();
+            row["Name"] = name;
+            row["Value"] = value;
+            table.Rows.Add(row);
+            return row;
         }
     }
 }